hg: fix cacheID construction
The repo URL must be in the cache id.
The ref is purposely not included in it because we want to reuse the cached repository
when the ref moves.
And finally, we use a sha256 hash to mask any authentication data because we don't
want them to be readable in the cache folder name.

Signed-off-by: Christophe de Vienne <christophe.devienne@orus.io>